mini push

+++ Project/Project/Graphic/header/shaderCreater.h
@@ -19,7 +19,7 @@ public:
 
 	GLuint getShaderProgramID()const;
 	void createShader(std::string vertexShader, std::string geometryShader, std::string fragmentShader);
-	void setFloat(std::string name, int value)const;
+	void setFloat(std::string name, GLfloat value)const;
 	void setMat4(const std::string &name, const glm::mat4 &mat) const;
 };
